First Commit of GetSubtitle
The plan is to make a Python Script that will download an english
subtitle of a video and create a OS X service (Automator) based on the
script. The script is still unusable but can already obtain a video’s
unique hash and make a download request, missing the actual
downloading. The script was made using thesubd.com API.